Replace registering of translations at plugin install time to the plugin init phase
This change follows review comments in:
https://github.com/elastic/kibana/issues/6515#issuecomment-236237218
