+++ .gitignore
@@ -24,6 +24,7 @@ target
 .eslintcache
 /plugins/
 data
+installedPlugins
 disabledPlugins
 webpackstats.json
 config/kibana.dev.yml
