+++ src/optimize/index.js
@@ -1,22 +1,6 @@
-import fs from 'fs';
 import FsOptimizer from './fs_optimizer';
-import i18nPlugin from '../plugins/i18n/server/i18n/index';
-import kibanaPackage from '../utils/package_json';
-import Promise from 'bluebird';
-
-const readdir = Promise.promisify(fs.readdir);
-
 module.exports = async (kbnServer, server, config) => {
   if (!config.get('optimize.enabled')) return;
-
-  server.log(
-    ['info', 'i18n'],
-    `Registering core plugin translations. This may take a few minutes`
-  );
-  let start = Date.now();
-  await registerCoreTranslations();
-  let seconds = ((Date.now() - start) / 1000).toFixed(2);
-  server.log(['info', 'i18n'], `Registration of core plugin translations completed in ${seconds} seconds`);
 
   // the lazy optimizer sets up two threads, one is the server listening
   // on 5601 and the other is a server listening on 5602 that builds the
@@ -35,11 +19,13 @@ module.exports = async (kbnServer, server, config) => {
   server.exposeStaticDir('/bundles/{path*}', bundles.env.workingDir);
   await bundles.writeEntryFiles();
 
-  // in prod, only bundle when someing is missing or invalid
-  let invalidBundles = config.get('optimize.useBundleCache') ? await bundles.getInvalidBundles() : bundles;
+  // in prod, only bundle what looks invalid or missing
+  if (config.get('optimize.useBundleCache')) {
+    bundles = await bundles.getInvalidBundles();
+  }
 
   // we might not have any work to do
-  if (!invalidBundles.getIds().length) {
+  if (!bundles.getIds().length) {
     server.log(
       ['debug', 'optimize'],
       `All bundles are cached and ready to go!`
@@ -62,27 +48,9 @@ module.exports = async (kbnServer, server, config) => {
     `Optimizing and caching ${bundles.desc()}. This may take a few minutes`
   );
 
-  start = Date.now();
+  let start = Date.now();
   await optimizer.run();
-  seconds = ((Date.now() - start) / 1000).toFixed(2);
+  let seconds = ((Date.now() - start) / 1000).toFixed(2);
 
   server.log(['info', 'optimize'], `Optimization of ${bundles.desc()} complete in ${seconds} seconds`);
-
 };
-
-function registerCoreTranslations()
-{
-  const rootDir = kibanaPackage.__dirname;
-
-  //Add translation dirs for the core plugins here
-  const corePluginTranslationDirs = [rootDir + '/src/ui/i18n'];
-
-  return Promise.map(corePluginTranslationDirs, (dir) => {
-    readdir(dir).then((dirListing) => {
-      Promise.map(dirListing, (listing) => {
-        const fullFilePath = dir + '/' + listing;
-        i18nPlugin.registerTranslations(fullFilePath);
-      });
-    });
-  });
-}
