+++ src/plugins/i18n/server/__tests__/fixtures/translations/test_plugin_1/en.json
@@ -1,0 +1,6 @@
+{
+  "test_plugin_1-NO_SSL": "Dont run the dev server using HTTPS",
+  "test_plugin_1-DEV": "Run the server with development mode defaults",
+  "test_plugin_1-NO_RUN_SERVER": "Dont run the dev server",
+  "test_plugin_1-HOME": "Run along home now!"
+}
