+++ src/plugins/i18n/server/__tests__/fixtures/translations/test_plugin_2/en.json
@@ -1,0 +1,6 @@
+{
+  "test_plugin_2-XXXXXX": "This is XXXXXX string",
+  "test_plugin_2-YYYY_PPPP": "This is YYYY_PPPP string",
+  "test_plugin_2-FFFFFFFFFFFF": "This is FFFFFFFFFFFF string",
+  "test_plugin_2-ZZZ": "This is ZZZ string"
+}
