--- src/ui/i18n/en-US.json
@@ -1,4 +1,0 @@
-{
-  "CORE-WELCOME_MESSAGE": "is loading. Give me a moment here. I'm loading a whole bunch of code. Don't worry, all this good stuff will be cached up for next time!",
-  "CORE-WELCOME_ERROR": "Kibana did not load properly. Check the server output for more information."
-}
